+++ lib/src/arg_parser.dart
@@ -4,6 +4,7 @@
 
 import 'dart:collection';
 
+import 'allow_anything_parser.dart';
 import 'arg_results.dart';
 import 'option.dart';
 import 'parser.dart';
@@ -29,6 +30,10 @@ class ArgParser {
   /// arguments.
   final bool allowTrailingOptions;
 
+  /// Whether or not this parser treats unrecognized options as non-option
+  /// arguments.
+  bool get allowsAnything => false;
+
   /// Creates a new ArgParser.
   ///
   /// If [allowTrailingOptions] is `true` (the default), the parser will parse
@@ -39,15 +44,22 @@ class ArgParser {
       <String, Option>{}, <String, ArgParser>{},
       allowTrailingOptions: allowTrailingOptions);
 
+  /// Creates a new ArgParser that treats *all input* as non-option arguments.
+  ///
+  /// This is intended to allow arguments to be passed through to child
+  /// processes without needing to be redefined in the parent.
+  ///
+  /// Options may not be defined for this parser.
+  factory ArgParser.allowAnything() = AllowAnythingParser;
+
   ArgParser._(Map<String, Option> options, Map<String, ArgParser> commands,
       {bool allowTrailingOptions: true})
       : this._options = options,
         this.options = new UnmodifiableMapView(options),
         this._commands = commands,
         this.commands = new UnmodifiableMapView(commands),
-        this.allowTrailingOptions = allowTrailingOptions != null
-            ? allowTrailingOptions
-            : false;
+        this.allowTrailingOptions =
+            allowTrailingOptions != null ? allowTrailingOptions : false;
 
   /// Defines a command.
   ///
@@ -69,10 +81,16 @@ class ArgParser {
   ///
   /// * There is already an option named [name].
   /// * There is already an option using abbreviation [abbr].
-  void addFlag(String name, {String abbr, String help, bool defaultsTo: false,
-      bool negatable: true, void callback(bool value), bool hide: false}) {
+  void addFlag(String name,
+      {String abbr,
+      String help,
+      bool defaultsTo: false,
+      bool negatable: true,
+      void callback(bool value),
+      bool hide: false}) {
     _addOption(name, abbr, help, null, null, null, defaultsTo, callback,
-        OptionType.FLAG, negatable: negatable, hide: hide);
+        OptionType.FLAG,
+        negatable: negatable, hide: hide);
   }
 
   /// Defines a value-taking option. Throws an [ArgumentError] if:
@@ -80,9 +98,16 @@ class ArgParser {
   /// * There is already an option with name [name].
   /// * There is already an option using abbreviation [abbr].
   /// * [splitCommas] is passed but [allowMultiple] is `false`.
-  void addOption(String name, {String abbr, String help, String valueHelp,
-      List<String> allowed, Map<String, String> allowedHelp, String defaultsTo,
-      void callback(value), bool allowMultiple: false, bool splitCommas,
+  void addOption(String name,
+      {String abbr,
+      String help,
+      String valueHelp,
+      List<String> allowed,
+      Map<String, String> allowedHelp,
+      String defaultsTo,
+      void callback(value),
+      bool allowMultiple: false,
+      bool splitCommas,
       bool hide: false}) {
     if (!allowMultiple && splitCommas != null) {
       throw new ArgumentError(
@@ -94,10 +119,19 @@ class ArgParser {
         splitCommas: splitCommas, hide: hide);
   }
 
-  void _addOption(String name, String abbr, String help, String valueHelp,
-      List<String> allowed, Map<String, String> allowedHelp, defaultsTo,
-      void callback(value), OptionType type,
-      {bool negatable: false, bool splitCommas, bool hide: false}) {
+  void _addOption(
+      String name,
+      String abbr,
+      String help,
+      String valueHelp,
+      List<String> allowed,
+      Map<String, String> allowedHelp,
+      defaultsTo,
+      void callback(value),
+      OptionType type,
+      {bool negatable: false,
+      bool splitCommas,
+      bool hide: false}) {
     // Make sure the name isn't in use.
     if (_options.containsKey(name)) {
       throw new ArgumentError('Duplicate option "$name".');
@@ -112,8 +146,8 @@ class ArgParser {
       }
     }
 
-    var option = newOption(name, abbr, help, valueHelp, allowed,
-        allowedHelp, defaultsTo, callback, type,
+    var option = newOption(name, abbr, help, valueHelp, allowed, allowedHelp,
+        defaultsTo, callback, type,
         negatable: negatable, splitCommas: splitCommas, hide: hide);
     _options[name] = option;
     _optionsAndSeparators.add(option);