+++ lib/src/arg_parser_exception.dart
@@ -2,7 +2,7 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-/// An exception thrown by [ArgParser].
+/// An exception thrown by `ArgParser`.
 class ArgParserException extends FormatException {
   /// The command(s) that were parsed before discovering the error.
   ///
@@ -10,8 +10,7 @@ class ArgParserException extends FormatException {
   final List<String> commands;
 
   ArgParserException(String message, [Iterable<String> commands])
-      : commands = commands == null
-            ? const []
-            : new List.unmodifiable(commands),
+      : commands =
+            commands == null ? const [] : new List.unmodifiable(commands),
         super(message);
 }