+++ lib/src/parser.dart
@@ -7,9 +7,9 @@ import 'arg_parser_exception.dart';
 import 'arg_results.dart';
 import 'option.dart';
 
-final _SOLO_OPT = new RegExp(r'^-([a-zA-Z0-9])$');
-final _ABBR_OPT = new RegExp(r'^-([a-zA-Z0-9]+)(.*)$');
-final _LONG_OPT = new RegExp(r'^--([a-zA-Z\-_0-9]+)(=(.*))?$');
+final _soloOpt = new RegExp(r'^-([a-zA-Z0-9])$');
+final _abbrOpt = new RegExp(r'^-([a-zA-Z0-9]+)(.*)$');
+final _longOpt = new RegExp(r'^--([a-zA-Z\-_0-9]+)(=(.*))?$');
 
 /// The actual argument parsing class.
 ///
@@ -47,7 +47,12 @@ class Parser {
   /// Parses the arguments. This can only be called once.
   ArgResults parse() {
     var arguments = args.toList();
-    var commandResults = null;
+    if (grammar.allowsAnything) {
+      return newArgResults(
+          grammar, const {}, commandName, null, arguments, arguments);
+    }
+
+    ArgResults commandResults;
 
     // Parse the args.
     while (args.length > 0) {
@@ -70,8 +75,7 @@ class Parser {
         } on ArgParserException catch (error) {
           if (commandName == null) rethrow;
           throw new ArgParserException(
-              error.message,
-              [commandName]..addAll(error.commands));
+              error.message, [commandName]..addAll(error.commands));
         }
 
         // All remaining arguments were passed to command so clear them here.
@@ -121,7 +125,7 @@ class Parser {
   /// We treat this differently than collapsed abbreviations (like "-abc") to
   /// handle the possible value that may follow it.
   bool parseSoloOption() {
-    var soloOpt = _SOLO_OPT.firstMatch(current);
+    var soloOpt = _soloOpt.firstMatch(current);
     if (soloOpt == null) return false;
 
     var option = grammar.findByAbbreviation(soloOpt[1]);
@@ -147,7 +151,7 @@ class Parser {
   /// (like "-abc") or a single abbreviation with the value directly attached
   /// to it (like "-mrelease").
   bool parseAbbreviation(Parser innermostCommand) {
-    var abbrOpt = _ABBR_OPT.firstMatch(current);
+    var abbrOpt = _abbrOpt.firstMatch(current);
     if (abbrOpt == null) return false;
 
     // If the first character is the abbreviation for a non-flag option, then
@@ -167,7 +171,8 @@ class Parser {
     } else {
       // If we got some non-flag characters, then it must be a value, but
       // if we got here, it's a flag, which is wrong.
-      validate(abbrOpt[2] == '',
+      validate(
+          abbrOpt[2] == '',
           'Option "-$c" is a flag and cannot handle value '
           '"${abbrOpt[1].substring(1)}${abbrOpt[2]}".');
 
@@ -206,7 +211,7 @@ class Parser {
   /// Tries to parse the current argument as a long-form named option, which
   /// may include a value like "--mode=release" or "--mode release".
   bool parseLongOption() {
-    var longOpt = _LONG_OPT.firstMatch(current);
+    var longOpt = _longOpt.firstMatch(current);
     if (longOpt == null) return false;
 
     var name = longOpt[1];