+++ test/usage_test.dart
@@ -11,7 +11,9 @@ void main() {
       var parser = new ArgParser();
       parser.addFlag('mode', help: 'The mode');
 
-      validateUsage(parser, '''
+      validateUsage(
+          parser,
+          '''
           --[no-]mode    The mode
           ''');
     });
@@ -20,7 +22,9 @@ void main() {
       var parser = new ArgParser();
       parser.addFlag('mode', negatable: false, help: 'The mode');
 
-      validateUsage(parser, '''
+      validateUsage(
+          parser,
+          '''
           --mode    The mode
           ''');
     });
@@ -29,7 +33,9 @@ void main() {
       var parser = new ArgParser();
       parser.addFlag('mode', help: 'The mode');
 
-      validateUsage(parser, '''
+      validateUsage(
+          parser,
+          '''
           --[no-]mode    The mode
           ''');
     });
@@ -39,7 +45,9 @@ void main() {
       parser.addFlag('mode', abbr: 'm', help: 'The mode');
       parser.addOption('long', help: 'Lacks an abbreviation');
 
-      validateUsage(parser, '''
+      validateUsage(
+          parser,
+          '''
           -m, --[no-]mode    The mode
               --long         Lacks an abbreviation
           ''');
@@ -50,7 +58,9 @@ void main() {
       parser.addFlag('mode', abbr: 'm', help: 'Lined up with below');
       parser.addOption('a-really-long-name', help: 'Its help text');
 
-      validateUsage(parser, '''
+      validateUsage(
+          parser,
+          '''
           -m, --[no-]mode             Lined up with below
               --a-really-long-name    Its help text
           ''');
@@ -60,7 +70,9 @@ void main() {
       var parser = new ArgParser();
       parser.addFlag('mode', help: '\n\n\n\nAfter newlines');
 
-      validateUsage(parser, '''
+      validateUsage(
+          parser,
+          '''
           --[no-]mode    After newlines
           ''');
     });
@@ -69,7 +81,9 @@ void main() {
       var parser = new ArgParser();
       parser.addFlag('mode', help: 'Before newlines\n\n\n\n');
 
-      validateUsage(parser, '''
+      validateUsage(
+          parser,
+          '''
           --[no-]mode    Before newlines
           ''');
     });
@@ -80,7 +94,9 @@ void main() {
       parser.addFlag('monkey', help: 'Second');
       parser.addFlag('wombat', help: 'Third');
 
-      validateUsage(parser, '''
+      validateUsage(
+          parser,
+          '''
           --[no-]zebra     First
           --[no-]monkey    Second
           --[no-]wombat    Third
@@ -92,7 +108,9 @@ void main() {
       parser.addFlag('affirm', help: 'Should be on', defaultsTo: true);
       parser.addFlag('negate', help: 'Should be off', defaultsTo: false);
 
-      validateUsage(parser, '''
+      validateUsage(
+          parser,
+          '''
           --[no-]affirm    Should be on
                            (defaults to on)
 
@@ -104,7 +122,9 @@ void main() {
       var parser = new ArgParser();
       parser.addOption('any', help: 'Can be anything', defaultsTo: 'whatevs');
 
-      validateUsage(parser, '''
+      validateUsage(
+          parser,
+          '''
           --any    Can be anything
                    (defaults to "whatevs")
           ''');
@@ -115,7 +135,9 @@ void main() {
       parser.addOption('out',
           abbr: 'o', help: 'Where to write file', valueHelp: 'path');
 
-      validateUsage(parser, '''
+      validateUsage(
+          parser,
+          '''
           -o, --out=<path>    Where to write file
           ''');
     });
@@ -126,7 +148,9 @@ void main() {
           help: 'Like in cards',
           allowed: ['spades', 'clubs', 'hearts', 'diamonds']);
 
-      validateUsage(parser, '''
+      validateUsage(
+          parser,
+          '''
           --suit    Like in cards
                     [spades, clubs, hearts, diamonds]
           ''');
@@ -139,7 +163,9 @@ void main() {
           defaultsTo: 'clubs',
           allowed: ['spades', 'clubs', 'hearts', 'diamonds']);
 
-      validateUsage(parser, '''
+      validateUsage(
+          parser,
+          '''
           --suit    Like in cards
                     [spades, clubs (default), hearts, diamonds]
           ''');
@@ -150,15 +176,22 @@ void main() {
       parser.addOption('suit',
           help: 'Like in cards',
           defaultsTo: 'clubs',
-          allowed: ['spades', 'clubs', 'diamonds', 'hearts'],
+          allowed: [
+            'spades',
+            'clubs',
+            'diamonds',
+            'hearts'
+          ],
           allowedHelp: {
-        'spades': 'Swords of a soldier',
-        'clubs': 'Weapons of war',
-        'diamonds': 'Money for this art',
-        'hearts': 'The shape of my heart'
-      });
+            'spades': 'Swords of a soldier',
+            'clubs': 'Weapons of war',
+            'diamonds': 'Money for this art',
+            'hearts': 'The shape of my heart'
+          });
 
-      validateUsage(parser, '''
+      validateUsage(
+          parser,
+          '''
           --suit              Like in cards
 
                 [clubs]       Weapons of war
@@ -174,7 +207,9 @@ void main() {
       parser.addOption('second', hide: true);
       parser.addOption('third', help: 'The third option');
 
-      validateUsage(parser, '''
+      validateUsage(
+          parser,
+          '''
           --first    The first option
           --third    The third option
           ''');
@@ -186,7 +221,9 @@ void main() {
       parser.addFlag('second', hide: true);
       parser.addFlag('third', help: 'The third flag');
 
-      validateUsage(parser, '''
+      validateUsage(
+          parser,
+          '''
           --[no-]first    The first flag
           --[no-]third    The third flag
           ''');
@@ -198,7 +235,9 @@ void main() {
       parser.addFlag('second-very-long-option', hide: true);
       parser.addFlag('third', help: 'The third flag');
 
-      validateUsage(parser, '''
+      validateUsage(
+          parser,
+          '''
           --[no-]first    The first flag
           --[no-]third    The third flag
           ''');
@@ -213,7 +252,9 @@ void main() {
         parser.addSeparator('Marsupial:');
         parser.addFlag('wombat', help: 'Third');
 
-        validateUsage(parser, '''
+        validateUsage(
+            parser,
+            '''
             --[no-]zebra     First
 
             Primate:
@@ -230,7 +271,9 @@ void main() {
         parser.addSeparator('Primate:');
         parser.addFlag('monkey', help: 'Second');
 
-        validateUsage(parser, '''
+        validateUsage(
+            parser,
+            '''
             --[no-]zebra     Multi
                              line
 
@@ -244,7 +287,9 @@ void main() {
         parser.addSeparator('Equine:');
         parser.addFlag('zebra', help: 'First');
 
-        validateUsage(parser, '''
+        validateUsage(
+            parser,
+            '''
             Equine:
             --[no-]zebra    First
             ''');
@@ -255,7 +300,9 @@ void main() {
         parser.addFlag('zebra', help: 'First');
         parser.addSeparator('Primate:');
 
-        validateUsage(parser, '''
+        validateUsage(
+            parser,
+            '''
             --[no-]zebra    First
 
             Primate:
@@ -267,7 +314,9 @@ void main() {
         parser.addSeparator('First');
         parser.addSeparator('Second');
 
-        validateUsage(parser, '''
+        validateUsage(
+            parser,
+            '''
             First
 
             Second