+++ test/utils.dart
@@ -9,6 +9,7 @@ import 'package:args/command_runner.dart';
 import 'package:test/test.dart';
 
 class CommandRunnerWithFooter extends CommandRunner {
+  @override
   String get usageFooter => "Also, footer!";
 
   CommandRunnerWithFooter(String executableName, String description)
@@ -18,55 +19,88 @@ class CommandRunnerWithFooter extends CommandRunner {
 class FooCommand extends Command {
   var hasRun = false;
 
+  @override
   final name = "foo";
+
+  @override
   final description = "Set a value.";
+
+  @override
   final takesArguments = false;
 
+  @override
   void run() {
     hasRun = true;
   }
 }
 
 class ValueCommand extends Command<int> {
+  @override
   final name = "foo";
+
+  @override
   final description = "Return a value.";
+
+  @override
   final takesArguments = false;
 
+  @override
   int run() => 12;
 }
 
 class AsyncValueCommand extends Command<String> {
+  @override
   final name = "foo";
+
+  @override
   final description = "Return a future.";
+
+  @override
   final takesArguments = false;
 
+  @override
   Future<String> run() async => "hi";
 }
 
 class MultilineCommand extends Command {
   var hasRun = false;
 
+  @override
   final name = "multiline";
+
+  @override
   final description = "Multi\nline.";
+
+  @override
   final takesArguments = false;
 
+  @override
   void run() {
     hasRun = true;
   }
 }
 
 class MultilineSummaryCommand extends MultilineCommand {
+  @override
   String get summary => description;
 }
 
 class HiddenCommand extends Command {
   var hasRun = false;
 
+  @override
   final name = "hidden";
+
+  @override
   final description = "Set a value.";
+
+  @override
   final hidden = true;
+
+  @override
   final takesArguments = false;
 
+  @override
   void run() {
     hasRun = true;
   }
@@ -75,11 +109,19 @@ class HiddenCommand extends Command {
 class AliasedCommand extends Command {
   var hasRun = false;
 
+  @override
   final name = "aliased";
+
+  @override
   final description = "Set a value.";
+
+  @override
   final takesArguments = false;
+
+  @override
   final aliases = const ["alias", "als"];
 
+  @override
   void run() {
     hasRun = true;
   }
@@ -88,10 +130,16 @@ class AliasedCommand extends Command {
 class AsyncCommand extends Command {
   var hasRun = false;
 
+  @override
   final name = "async";
+
+  @override
   final description = "Set a value asynchronously.";
+
+  @override
   final takesArguments = false;
 
+  @override
   Future run() => new Future.value().then((_) => hasRun = true);
 }