+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 * **Breaking change**: The `allowTrailingOptions` argument to `new
   ArgumentParser()` defaults to `true` instead of `false`.
 
+* Add `new ArgParser.allowAnything()`. This allows any input, without parsing
+  any options.
+
 ## 0.13.7
 
 * Add explicit support for forwarding the value returned by `Command.run()` to